feat(shared): add Bold methodology homologation document helpers
