+++ libs/shared/methodologies/bold/helpers/.eslintrc.js
@@ -1,0 +1,5 @@
+const { getBaseEslintConfig } = require('../../../../../.eslint/eslint.config');
+
+module.exports = getBaseEslintConfig({
+  projectPath: __dirname,
+});
