+++ libs/shared/methodologies/bold/helpers/jest.config.ts
@@ -1,0 +1,3 @@
+import { getJestBaseConfig } from '../../../../../.jest/config/jest.base.config';
+
+export default getJestBaseConfig(__dirname);
