+++ libs/shared/methodologies/bold/helpers/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@carrot-fndn/shared/methodologies/bold/helpers",
+  "version": "0.0.0"
+}
