+++ libs/shared/methodologies/bold/helpers/project.json
@@ -1,0 +1,12 @@
+{
+  "name": "shared-methodologies-bold-helpers",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "tags": ["type:lib", "app:any", "stack:any"],
+  "sourceRoot": "libs/shared/methodologies/bold/helpers/src",
+  "projectType": "library",
+  "targets": {
+    "lint": {},
+    "test": {},
+    "ts": {}
+  }
+}
