+++ libs/shared/methodologies/bold/helpers/src/index.ts
@@ -1,0 +1,1 @@
+export * from './homologation-document.helpers';
