+++ libs/shared/methodologies/bold/helpers/tsconfig.build.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["./tsconfig.json", "../../../.tsconfig/tsconfig.build.json"]
+}
