+++ libs/shared/methodologies/bold/helpers/tsconfig.spec.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../../../.tsconfig/tsconfig.spec.json"
+}
