+++ libs/shared/methodologies/bold/processors/src/geolocation-precision.helpers.ts
@@ -8,23 +8,15 @@ import {
   PARTICIPANT_HOMOLOGATION_PARTIAL_MATCH,
 } from '@carrot-fndn/shared/methodologies/bold/matchers';
 import {
-  eventNameIsAnyOf,
-  isOpenEvent,
-} from '@carrot-fndn/shared/methodologies/bold/predicates';
-import {
-  type Document,
   type DocumentEvent,
   DocumentEventAttributeName,
-  DocumentEventName,
   type DocumentSubtype,
 } from '@carrot-fndn/shared/methodologies/bold/types';
 import {
   type Latitude,
   type Longitude,
   type MethodologyAddress,
-  type NonEmptyString,
 } from '@carrot-fndn/shared/types';
-import { isAfter, startOfToday } from 'date-fns';
 import { is } from 'typia';
 
 export const participantHomologationCriteria = (
@@ -50,37 +42,6 @@ export const participantHomologationCriteria = (
     },
   ],
 });
-
-export const getParticipantHomologationDocument = ({
-  homologationDocuments,
-  participantId,
-}: {
-  homologationDocuments: Document[];
-  participantId: NonEmptyString;
-}): Document | undefined =>
-  homologationDocuments.find((document) => {
-    const openEvent = document.externalEvents?.find(isOpenEvent);
-
-    return !isNil(openEvent) && openEvent.participant.id === participantId;
-  });
-
-export const homologationIsNotExpired = (document: Document): boolean => {
-  const closeEvent = document.externalEvents?.find(
-    eventNameIsAnyOf([DocumentEventName.CLOSE]),
-  );
-  const homologationDueDate = getEventAttributeValue(
-    closeEvent,
-    DocumentEventAttributeName.HOMOLOGATION_DUE_DATE,
-  );
-
-  if (is<NonEmptyString>(homologationDueDate)) {
-    const dueDate = new Date(homologationDueDate);
-
-    return isAfter(dueDate, startOfToday());
-  }
-
-  return false;
-};
 
 interface Geolocation {
   latitude: number;
